refactor(parser-support): share oracle validation modules
