+++ components/aihc-parser-support/aihc-parser-support.cabal
@@ -1,0 +1,31 @@
+cabal-version:      3.8
+name:               aihc-parser-support
+version:            0.1.0.0
+build-type:         Simple
+license:            MIT
+license-file:       ../../LICENSE
+author:             David Himmelstrup
+maintainer:         lemmih@gmail.com
+category:           Compilers
+synopsis:           Shared parser validation and oracle support for aihc
+
+library
+  exposed-modules:
+      CppSupport
+    , GhcOracle
+    , ParserValidation
+  hs-source-dirs:   src
+  build-depends:
+      base >=4.16 && <5
+    , aihc-cpp
+    , aihc-hackage
+    , aihc-parser
+    , Diff
+    , bytestring
+    , deepseq
+    , filepath
+    , ghc-lib-parser
+    , prettyprinter
+    , text
+  ghc-options:        -Wall
+  default-language: GHC2021
