+++ tooling/aihc-dev/aihc-dev.cabal
@@ -9,40 +9,59 @@ maintainer:         lemmih@gmail.com
 category:           Development
 synopsis:           Developer tools for the aihc compiler
 
+library extract-hi
+  hs-source-dirs:   src
+  exposed-modules:
+      Aihc.Dev.ExtractHi
+    , Aihc.Dev.ExtractHi.Types
+    , Aihc.Dev.ExtractHi.ToResolveIface
+  other-modules:
+      Aihc.Dev.ExtractHi.GhcSession
+  build-depends:
+      base >= 4.16 && < 5
+    , aeson >= 2.0 && < 2.3
+    , bytestring >= 0.10.8 && < 0.13
+    , containers >= 0.5 && < 0.8
+    , directory >= 1.2.3 && < 1.5
+    , filepath >= 1.3.0.1 && < 1.6
+    , ghc >= 9.12 && < 9.13
+    , process >= 1.6 && < 1.8
+    , text >= 1.2.3 && < 2.2
+  default-extensions: OverloadedStrings
+  ghc-options:        -Wall
+  default-language: GHC2021
+
+library snippet
+  hs-source-dirs:   src
+  exposed-modules:
+      Aihc.Dev.Snippet
+  build-depends:
+      base >= 4.16 && < 5
+    , aihc-cpp
+    , aihc-parser
+    , aihc-parser-support
+    , prettyprinter
+    , text >= 1.2.3 && < 2.2
+  default-extensions: OverloadedStrings
+  ghc-options:        -Wall
+  default-language: GHC2021
+
 executable aihc-dev
   main-is:          Main.hs
   hs-source-dirs:
       exe
-    , src
-  other-modules:
-       Aihc.Dev.ExtractHi
-     , Aihc.Dev.ExtractHi.Types
-     , Aihc.Dev.ExtractHi.GhcSession
-     , Aihc.Dev.ExtractHi.ToResolveIface
-     , Aihc.Dev.Snippet
-     , Aihc.Dev.Snippet.CppSupport
-     , Aihc.Dev.Snippet.GhcOracle
-     , Aihc.Dev.Snippet.Validation
   build-depends:
-       base >= 4.16 && < 5
-     , aeson >= 2.0 && < 2.3
-     , aeson-pretty >= 0.8 && < 0.9
-     , aihc-cpp
-     , aihc-hackage
-     , aihc-parser
-      , bytestring >= 0.10.8 && < 0.13
-      , containers >= 0.5 && < 0.8
-      , Diff
-      , deepseq
-     , directory >= 1.2.3 && < 1.5
-     , filepath >= 1.3.0.1 && < 1.6
-     , ghc >= 9.12 && < 9.13
-     , ghc-boot-th
-     , optparse-applicative >= 0.16 && < 0.19
-     , prettyprinter
-     , process >= 1.6 && < 1.8
-     , text >= 1.2.3 && < 2.2
-     , yaml >= 0.11 && < 0.12
+      base >= 4.16 && < 5
+    , aeson >= 2.0 && < 2.3
+    , aeson-pretty >= 0.8 && < 0.9
+    , aihc-parser
+    , aihc-dev:extract-hi
+    , aihc-dev:snippet
+    , bytestring >= 0.10.8 && < 0.13
+    , directory >= 1.2.3 && < 1.5
+    , filepath >= 1.3.0.1 && < 1.6
+    , optparse-applicative >= 0.16 && < 0.19
+    , yaml >= 0.11 && < 0.12
   default-extensions: OverloadedStrings
   ghc-options:        -Wall
   default-language: GHC2021
@@ -51,25 +70,11 @@ test-suite spec
   type:             exitcode-stdio-1.0
   hs-source-dirs:
       test
-    , src
   main-is:          Spec.hs
-  other-modules:
-      Aihc.Dev.Snippet
-    , Aihc.Dev.Snippet.CppSupport
-    , Aihc.Dev.Snippet.GhcOracle
-    , Aihc.Dev.Snippet.Validation
   build-depends:
       base >= 4.16 && < 5
-    , aihc-cpp
-    , aihc-hackage
     , aihc-parser
-    , bytestring >= 0.10.8 && < 0.13
-    , Diff
-    , deepseq
-    , filepath >= 1.3.0.1 && < 1.6
-    , ghc >= 9.12 && < 9.13
-    , ghc-boot-th
-    , prettyprinter
+    , aihc-dev:snippet
     , tasty
     , tasty-hunit
     , tasty-quickcheck
